+++ NEWS.md
@@ -13,6 +13,9 @@
 * Added support for logging using the `logger` package. Added a new function `register_logger`, which registers a logger in a given namespace.
 * Added support of the full screen for a `module` when the `filters` argument is equal `NULL`.
 
+### Bug fixes
+* Fixed data loading of `DatasetConnector` being dependent on other `Dataset` or `DatasetConnector` objects. 
+
 ### Miscellaneous
 * Replaced the servers from `DataConnection`, `RelationalDataConnector`, `DatasetConnector`, and `RelationalData` with `moduleServer`.
 * Updated R version requirement to >= 3.6.